2017.6.4 - 152656203 invoicecustomers changes

+++ KeithLink/KeithLink.Svc.Core/Models/Invoices/InvoiceCustomers.cs
@@ -20,7 +20,7 @@ namespace KeithLink.Svc.Core.Models.Invoices
         public int NumberCustomers { get; set; }
 
         [DataMember(Name = "totalnumberinvoices")]
-        public int TotalNumberInvoices { get; set; }
+        public long TotalNumberInvoices { get; set; }
 
         [DataMember(Name = "customers")]
         public List<InvoiceCustomer> customers { get; set; }
